requirements.txt
requests
pyyaml
bs4
pandas
boto3
lxml
